pyproject.toml
[tool.poetry]
name = "marketplace"
version = "0.1.0"
description = ""
authors = ["derijablya <dulugov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
black = "^24.1.1"
pydantic = "^2.6.1"
ruff = "^0.2.1"
setuptools = "^69.1.0"
wheel = "^0.42.0"
twine = "^5.0.0"
pytest = "^8.0.0"
pytest-mock = "^3.12.0"
requests-mock = "^1.11.0"
aiohttp = "^3.9.3"
aioresponses = "^0.7.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
